♻️ Rename component (#698)
